requirements.txt
attrs==23.2.0
blinker==1.7.0
cachelib==0.10.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
country-converter==1.2
fiona==1.9.5
Flask==3.0.1
Flask-Session==0.6.0
geographiclib==2.0
geopandas==0.14.3
h11==0.14.0
haversine==2.8.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
numpy==1.26.3
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.0
psycopg2-binary==2.9.9
pyproj==3.6.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.4
requests==2.31.0
selenium==4.17.2
setuptools==69.0.3
shapely==2.0.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
Werkzeug==3.0.1
wsproto==1.2.0